working on the attendance information

+++ src/pages/Lecturer/AttendanceList.jsx
@@ -76,7 +76,7 @@ export default function AttendanceList() {
                           "en-US",
                           {
                             year: "numeric",
-                            month: "long",
+                            month: "numeric",
                             day: "numeric",
                           }
                         )}
